upatre

+++ src/app/admin_general_component/penalty_kickout.tsx
@@ -3,24 +3,54 @@
 import { initializeApp } from "firebase/app";
 import { useRouter } from "next/navigation";
 import firebaseConfig from "../utils/fire_base_config";
-import { collection, deleteDoc, doc, getFirestore } from "firebase/firestore";
+import {
+  collection,
+  deleteDoc,
+  doc,
+  getDoc,
+  getDocs,
+  getFirestore,
+} from "firebase/firestore";
+import { cancelSubscription } from "../utils/stripe";
+import { useState } from "react";
 
 const Penalty_kickout = ({ sethide_penalty_kickout, user_doc_id }: any) => {
   const router = useRouter();
+  const [btn_text, setbtn_text] = useState("Kick out");
 
   const app = initializeApp(firebaseConfig);
 
   // Initialize Firestore
   const db = getFirestore(app);
-  const deleteuser = () => {
+  const deleteuser = async () => {
     const user_ref = doc(db, "users", user_doc_id);
-    deleteDoc(user_ref)
-      .then(() => {
-        window?.location?.reload();
-      })
-      .catch((err) => {
-        console.log("Error deleting user" + err);
-      });
+    setbtn_text("Kick out");
+    try {
+      const userDocSnapshot = await getDoc(user_ref);
+      setbtn_text("Processing");
+      if (userDocSnapshot.exists()) {
+        const userData = userDocSnapshot.data();
+        // console.log("User data:", userData.subscriptionId);
+
+        userData.subscriptionId &&
+          (await cancelSubscription(userData.subscriptionId));
+
+        deleteDoc(user_ref)
+          .then(() => {
+            window?.location?.reload();
+          })
+          .catch((err) => {
+            console.log("Error deleting user" + err);
+          });
+      } else {
+        console.log("User not found");
+        setbtn_text("No user");
+      }
+    } catch (error) {
+      console.error("Error getting user document:", error);
+      setbtn_text("Error");
+    }
+
     // router.replace();
   };
   return (
@@ -48,7 +78,7 @@ const Penalty_kickout = ({ sethide_penalty_kickout, user_doc_id }: any) => {
                 deleteuser();
               }}
             >
-              Kick out
+              {btn_text}
             </button>
             <button
               className="h-[3.5vw] sm:h-[13vw] sm:text-[3.5vw] w-full bg-[#CCFF00] neuer text-black hover:bg-opacity-[80%] text-[0.9vw] rounded-[1vw] sm:rounded-[3vw]"
